pyproject.toml
[project]
name = "Haruki-Sekai-API"
version = "4.0.0"
description = "An unofficial API proxy project for Project Sekai"
requires-python = ">=3.11"
dependencies = [
    "aiofiles~=23.2.1",
    "aiohttp~=3.11.12",
    "aiopath>=0.6.11",
    "apscheduler~=3.10.4",
    "coloredlogs~=15.0.1",
    "cryptography~=44.0.1",
    "fastapi>=0.115.12",
    "hypercorn~=0.16.0",
    "msgpack~=1.0.8",
    "orjson>=3.10.18",
    "pydantic~=2.10.6",
    "pygit2~=1.17.0",
    "pyjwt>=2.10.1",
    "redis>=6.2.0",
    "ruff>=0.11.9",
    "sqlalchemy>=2.0.41",
    "tenacity~=9.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "W", # pycodestyle warnings
    "E", # pycodestyle errors
    "I", # isort
    "UP", # pyupgrade
    "ASYNC", # flake8-async
    "C4", # flake8-comprehensions
    "T10", # flake8-debugger
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "TID", # flake8-tidy-imports
    "RUF", # Ruff-specific rules
]
ignore = [
    "E402", # module-import-not-at-top-of-file
    "UP037", # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["nonebot", "tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
